fix: issue #356
